Change translation library
react-intl was causing a complete render of the whole component tree
whenever the language was changing. The only workaround I found was to
remove the `key` parameter from the Provider, but it also caused
side-effects.

Elements were not rendering properly, and render had to be triggered
connecting the language key from the store, which is not satisfactory.

i18next approach prevents this to happen and doesn't seem to be less
performant.