Improve (un)mute and volume handling
When using the player native controls, we must ensure state will be
updated. But when we update our state with custom controls, the video
element properties must be updated too.
When a property is shared between the HTMLVideoElement and the state, we
need to reflect local element changes to the state. This leads to
multiple event calling, or worse, infinite dispatch loop.

To prevent this, the video element will track its own changes and
resolve differences with application state on its own.