Change tracks display policy
When using non native controls, video tracks have to be hidden from the
users, even though we still need to use the cue changes.

For the moment, reactivating text tracks to use in a native fashion is
not possible, but could be achieved by managing each tracks in the
application state.