Use unique id HOC, refactor to function components
Some DOM elements need an `id` to be made accessible.
For example when using the `aria-labelledby` attribute.

Generating the id in the render method is not optimal since a new id
will be generated every time an element will be rendered. This is why
the id generation was moved in a higher order component.

React `SFC` and `StatelessComponent` types are being deprecated in
favor of `FunctionComponent`. This is mostly a naming convention and
follows the introduction of React Hooks, but it implied changing all
component types in the codebase.
Bare functions are recognized as `FunctionComponent`, can be used in
most situations, and are also more readable (they are "just" functions).
This is why all simple components were changed into bare functions, and
this is the way these components should be written from now.

Using the `const` keyword saves very few lines of code and brings no
benefit whatsoever.