Hide controls after a period of inactivity
When user is not performing any interaction with the player, I think it
is safe to assume some controls can be visually hidden from the GUI.

This commit adds this feature by listening to a list of events on the
main element. After a default of 3000ms, the controls and the timeline
will hide themselves. Any event from the list will show them again.