Remove windows boundaries options
Specifying lower and upper bounds for a window may seem like a good idea
(or at least not a bad one), but raises a few questions.
The most important one is: "Why would we constrain the movement of a
window even more?".
Since I don't have the answer and cannot say if this is relevant or not,
I assume an element is free to move between 0 and 100 percents of its
container width.