Change resizers display policy, update keyboard management
Drag button was displayed using static flow, resulting in windows being
enlarged by the component dimensions. To avoid this behavior, the drag
button is now displayed using absolute positioning.
It is shown when the button has the focus or when it is hovered by the
pointer device. These two states have different display styles to give
better visual feedback for keyboard users.

Also, buttons focus can be cancelled using the escape key.
This gives the user a way to hide the button when they are dragging or
resizing using the keyboard.