Fix i18n critical issue and upgrade packages
Interface language was always loading default language even if another
language was specified in loaded `preferences.currentLanguage`. Changing
language worked properly, the issue only concerned the initial loading.

Used react-i18next and i18next versions were outdated, and I couldn't
find anything related to my problems in docs or issues. Therefore I
decided it was the good time to do a complete upgrade of both packages.

A consequence of this upgrade was the need to use React v16.8 or higher,
since react-i18next in the most recent versions uses hooks, a feature
not supported in the version of react used in Aiana (16.6). React got
updated too.

For simple function components, there is no need to use the higher order
component provided by react-i18next anymore, the `useTranslation` hook
works just fine.
For class based components, a few renaming operations had to be
performed because the react-i18next API received a few changes.

In the process, I took the chance to remove translations from the
bundled application to use the backend utility and load them
asynchronously.
This allows updating the translations without having to ship a new
version of the complete application, and also allows the use of external
translation tools.