Allow locking windows, refactor state
To show or hide a window, a new property had to be set in the state.
The most suitable emplacement was inside dedicated preferences UI entry,
but its naming (uiPlacement) could have been confusing. To prevent any
confusion, it has been refactored to `uiWindows`.
Adding a `visible` property won't be misleading, and it should be added
in the next commits.

Since we could add properties, I took some time to implement the lock
feature, preventing a window to be closed, moved, or resized.

A toggle button was added to control the locked value.

Each window can be set to be locked independently from the other ones,
but the toggle button will toggle them all at once for the sake of
simplicity.
Should a more precise way of setting them be asked for, preferences will
be adapted to reflect this.