Add interactive transcript
Transcript is an accessibility requirement, and the media player was
lacking an option to provide one.

The transcript is a syndicated content made with chapters, subtitles,
audio descriptions and other time-based content. For the time being,
only chapters and subtitles are displayed, but adding new kind of
content such as additional informations is rather easy.