Allow using YouTube iframe as media
An abstraction layer was added to handle multiple types of medias.
However, a media type should provide functions to allow using any
external API, such as YouTube or other video provider.

Documentation is behind code at the moment, and using an abstract class
might be the best way to ensure methods are provided.

One consequence is the refactor of action creators to remove any kind of
mention to the media loaded by the player, and only modify the global
state. This is cleaner and avoids unexpected side effects.

The `seek` mechanism was also simplified, and fullscreen action was
turned into a util function.