Update global config and use dynamic container tag
`loggerEndpoint` has been moved to the `window.aiana` object to avoid
name collision with possibly present code.

Main container tag can be `main` or `section` depending on its context.
If it is used as an embed player in an iframe,
`window.aiana.isMainContent` should be set to `true`. If the player is
meant to be wrapped around existing HTML content (which may have a main
tag), then this option should not be set.